blog message end

+++ controllers/blogs.controller.js
@@ -21,20 +21,37 @@ exports.createBlogs = async (req, res, next) => {
     try {
 
         const body = req.body;
+        if (req.file) {
+            const {filename: image} = req.file;
+            await sharp(req.file.path)
+                .resize(800)
+                .webp({quality: 90})
+                .toFile(path.resolve(req.file.destination, "resized", image))
+            fs.unlinkSync(req.file.path);
 
-        const {filename: image} = req.file;
-        await sharp(req.file.path)
-            .resize(800)
-            .webp({quality: 90})
-            .toFile(path.resolve(req.file.destination, "resized", image))
-        fs.unlinkSync(req.file.path);
+            await createBlog({...body, image: req.file.filename, author: req.user._id, created: Date.now()})
+            const blogs = await findAllBlogs().populate('author');
+            res.render('admin/blogs/index', {blogs, success: `successfully added ${body.title}`, currentUser: req.user})
 
-        await createBlog({...body, image: req.file.filename, author: req.user._id, created: Date.now()})
-        res.redirect('/admin/blogs')
+        } else {
+
+            await createBlog({...body, author: req.user._id, created: Date.now()})
+            const blogs = await findAllBlogs().populate('author');
+            res.render('admin/blogs/index', {blogs, success: `successfully added ${body.title}`, currentUser: req.user})
+
+        }
+
 
     } catch (e) {
 
-        next(e)
+        let errors;
+        const blogs = await findAllBlogs().populate('author');
+        if (e.code) {
+            errors = ['duplicate key']
+        } else {
+            errors = Object.keys(e.errors).map(key => e.errors[key].message)
+        }
+        res.status(400).render('admin/blogs/index', {blogs, errors, currentUser: req.user})
 
     }
 }
@@ -44,12 +61,25 @@ exports.deleteBlogs = async (req, res, next) => {
     const blogID = req.params.id;
     const blog = await findBlogs(blogID);
 
-    const image = blog.image;
-    fs.unlink(path.join(__dirname, `../public/images/blogs/resized/${image}`), (err => err && console.error(err)))
-    await deleteBlog(blogID);
+    try {
 
-    const blogs = await findAllBlogs().populate('author')
-    res.render('admin/blogs/index', {blogs, currentUser: req.user})
+
+        let name = blog.title;
+
+        const image = blog.image;
+        fs.unlink(path.join(__dirname, `../public/images/blogs/resized/${image}`), (err => err && console.error(err)))
+
+        await deleteBlog(blogID);
+
+        const blogs = await findAllBlogs().populate('author')
+        res.render('admin/blogs/index', {blogs, success: `successfully deleted ${name}`, currentUser: req.user})
+
+    } catch (e) {
+
+        next(e)
+
+    }
+
 
 }
 
@@ -64,7 +94,7 @@ exports.getBlog = async (req, res, next) => {
 
     } catch (e) {
 
-        next(e)
+        res.render('404', {currentUser: req.user})
 
     }
 }
@@ -78,6 +108,7 @@ exports.updateBlogs = async (req, res, next) => {
         const body = req.body;
 
         if (req.file) {
+
             const blog = await findBlogs(blogID);
             const oldImage = blog.image;
             fs.unlink(path.join(__dirname, `../public/images/blogs/resized/${oldImage}`), (err => err && console.error(err)))
@@ -90,21 +121,38 @@ exports.updateBlogs = async (req, res, next) => {
                 .webp({quality: 90})
                 .toFile(path.resolve(req.file.destination, "resized", image))
             fs.unlinkSync(req.file.path);
-        }
 
-        await updateBlog(blogID, body);
-        res.redirect('/admin/blogs/' + blogID);
+            await updateBlog(blogID, body);
+            const blogs = await findAllBlogs().populate('author');
+            res.render('admin/blogs/index', {
+                blogs,
+                success: `successfully updated ${body.name}`,
+                currentUser: req.user
+            });
+
+        } else {
+
+            await updateBlog(blogID, body);
+            const blogs = await findAllBlogs().populate('author');
+            res.render('admin/blogs/index', {
+                blogs,
+                success: `successfully updated ${body.name}`,
+                currentUser: req.user
+            });
+
+        }
 
     } catch (e) {
 
-        const errors = Object.keys(e.errors).map(key => e.errors[key].message)
-        const blog = await findBlogs(blogID);
-        res.status(400).render('/blog-form', {
-            errors,
-            blog,
-            isAuthenticated: req.isAuthenticated(),
-            currentUser: req.user
-        });
+        let errors;
+        const blog = await findBlogs(blogID).populate('author');
+
+        if (e.code) {
+            errors = ['duplicate key']
+        } else {
+            errors = Object.keys(e.errors).map(key => e.errors[key].message)
+        }
+        res.status(400).render('admin/blogs/blog', {blog, errors, currentUser: req.user})
 
     }
 }
